Add gn:traitId and gn:publicationId.
In GeneNetwork, a phenonytpe is currently identified by it's
ID (primary key of the table from MariaDB). The only way to relate it
to a publication is through a publication ID. This is important
because there are some publications with a NULL value for "pubmed ID"
and as such without the publication ID, some data is lost as there's
no way to point to publication with a NULL "pubmed ID."

* dump.scm (dump-publish-xref): Define gn:traitId and
gn:publicationId.

Signed-off-by: Arun Isaac <arunisaac@systemreboot.net>